+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Changed simple value serialization to distinguish an empty object/array with non-empty
 - Added simple value changes description for hooks
 - Fixed an order of rerender events
+- Changed parent update -> owner update and improved its displaying in event log
 
 ## 0.2.0 (August 27, 2021)